backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.2
SQLAlchemy==2.0.29
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==25.1.0